changed layout of story title and renamed it to story image. The title text was moved to the article section.
